Extract command module
