+++ src/sode/python/sode/cli/command.py
@@ -1,0 +1,15 @@
+import io
+from abc import abstractmethod
+from dataclasses import dataclass
+
+
+@dataclass
+class RunState:
+    stderr: io.TextIOBase
+    stdout: io.TextIOBase
+
+
+class CliCommand:
+    @abstractmethod
+    def run(self, state: RunState) -> int:
+        pass
